Finished uno gam
Not sure how well this is reflected in the github as I pretty much have been just working on and off from in my hercules account for the past few days because it's easier to test there. Anyways here is the pretty much finished product. The game starts when 4 clients connect to the server then player 1 needs to press enter to receive the prompt for the first turn. The game will tell the player if the player is able to take a turn (i.e. if the turn has been skipped, if the player has to draw 2/4 or if the player has a card that is currently playable. The game simulates a deck by comparing the cards that are currently in play + the cards that have been played and referencing that with an array representing the possible cards in the deck, once all cards have been taken the deck gets "reshuffled" that is all 1 values become 0 aside from the cards that are in the player's hands. once a player has gotten rid of all their cards the game is over and the server closes.